xform: move join.opt functions to join_funcs.go
Release note: None
